Added an HTC Vive focused variant optimized for the Fluid dynamics domain. More information about this variant can be seen on Matthias Albert bachelor thesis: http://www.nm.ifi.lmu.de/pub/Fopras/albe18/
This is a fork from the NOMAD VR build for the HTC Vive from 2018.

It allows to view datasets with large amounts of isosurfaces.
To control how many timesteps are loaded a config file option is used:
isobuffersize [int]

If the dataset uses particles/atoms an interactive interface is shown above the second controller.
Above the first controller its x,y,z coordinates are displayed (current range from -9,99 to 9,99).
Additionally the velocity, acceleration and crossproduct (v x a) of all particles are shown.
To change their scaling the following config file options are used:
velocityscaling [float]
accelerationscaling [float]
crossproductscaling [float]